% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{get_plugins}
\alias{get_plugins}
\alias{get_core_plugins}
\alias{get_community_plugins}
\title{Get Plugins}
\usage{
get_plugins(vault_path, ...)

get_core_plugins(vault_path, ...)

get_community_plugins(vault_path, ...)
}
\arguments{
\item{vault_path}{Path to the Obsidian Vault's root-level directory.}

\item{...}{Not used}
}
\value{
Returns a list of character vectors for both "core" and "community"
plugins.
}
\description{
Retrieve details regarding an Obsidian Vault's utilized plugins, both core
and community.
}
\section{Functions}{
\itemize{
\item \code{get_core_plugins}: Get Core Obsidian Plugins

\item \code{get_community_plugins}: Get Community Obsidian Plugins
}}

\examples{
\dontrun{

 # get plugins from a vault
 vault_path <- "path/to/my/vault"
 get_plugins(vault_path)

}
}
