% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itinerary_hawaii_m.R
\name{mod_itinerary_m_server}
\alias{mod_itinerary_m_server}
\title{Server du module navigateur d'itinéraire (mobile)}
\usage{
mod_itinerary_m_server(id, voyage_data)
}
\arguments{
\item{id}{ID du module Shiny (doit correspondre à `mod_itinerary_m_ui`)}

\item{voyage_data}{Liste ou reactive contenant les données du voyage.}
}
\description{
Server du module navigateur d'itinéraire (mobile)
}
