% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itinerary_hawaii_m.R
\name{mod_itinerary_m_ui}
\alias{mod_itinerary_m_ui}
\title{UI du module navigateur d'itinéraire (mobile)}
\usage{
mod_itinerary_m_ui(id)
}
\arguments{
\item{id}{ID du module Shiny}
}
\description{
UI du module navigateur d'itinéraire (mobile)
}
