% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itinerary_hawaii.R
\name{mod_itinerary_server}
\alias{mod_itinerary_server}
\title{Server du module navigateur d'itinéraire}
\usage{
mod_itinerary_server(id, voyage_data)
}
\arguments{
\item{id}{ID du module Shiny (doit correspondre à `mod_itinerary_ui`)}

\item{voyage_data}{Liste ou reactive contenant les données du voyage.
Produite par `build_voyage_hawaii.R` et chargée depuis S3.}
}
\description{
Server du module navigateur d'itinéraire
}
