% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_itinerary_hawaii.R
\name{mod_itinerary_ui}
\alias{mod_itinerary_ui}
\title{UI du module navigateur d'itinéraire}
\usage{
mod_itinerary_ui(id)
}
\arguments{
\item{id}{ID du module Shiny}
}
\value{
Un `tagList` bslib compatible, à insérer dans n'importe quelle
  `page_*` bslib (page_navbar, page_fluid, etc.).
}
\description{
UI du module navigateur d'itinéraire
}
