% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsidian-package.R
\docType{package}
\name{obsidian-package}
\alias{obsidian}
\alias{obsidian-package}
\title{obsidian: Analyze Obsidian Vaults from R}
\description{
Analyze Obsidian Vaults.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/obsidian}
  \item Report bugs at \url{https://github.com/jimbrig/obsidian/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
