% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ressource.R
\name{read_ressource}
\alias{read_ressource}
\title{Lire les notes ressource (livres, formations)}
\usage{
read_ressource(vault_path = NULL)
}
\arguments{
\item{vault_path}{Chemin vers le dossier vault. Si NULL, lu depuis config.yml.}
}
\value{
Un tibble avec une ligne par note : `title`, `auteur`, `tags`,
  `date_publication`, `body`, `path`.
}
\description{
Lit toutes les notes de type `note ressource` du vault.
}
