% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_helpers.R
\name{s3_list_vault}
\alias{s3_list_vault}
\title{Lister les fichiers du vault sur S3}
\usage{
s3_list_vault(bucket = NULL, prefix = NULL)
}
\arguments{
\item{bucket}{Nom du bucket S3. Si NULL, lu depuis config.yml.}

\item{prefix}{Préfixe (dossier) dans le bucket. Si NULL, lu depuis config.yml.}
}
\value{
Un tibble avec les clés S3 des fichiers du vault.
}
\description{
Lister les fichiers du vault sur S3
}
